Remove InitCommand trigger and fix binding typo
In `MainWindow.xaml`, removed the `<b:Interaction.Triggers>` section that triggered the `InitCommand` on window load, which may impact the initialization of the `MainViewModel`.

In `WebView.xaml`, corrected a typo in the binding path for the `Visibility` property of a `TextBlock` from `Reuqest.BodyType` to `Request.BodyType`. This ensures proper binding and improves the visual presentation of the `TextBlock` when the request body is empty.